pyproject.toml
[project]
name = "repo2doc"
version = "0.1.0"
description = "基于 LangGraph 的代码库逆向需求文档生成工具"
readme = "readme.cn.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "tiktoken>=0.8.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "pathspec>=0.12.0",
    "rich>=13.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nodes", "utils"]

[tool.ruff]
line-length = 100
target-version = "py311"
